requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.8.0
attrs==25.1.0
certifi==2024.12.14
click==8.1.8
fastapi==0.115.7
fastapi-utils==0.8.0
frozenlist==1.5.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
multidict==6.1.0
mypy-extensions==1.0.0
propcache==0.2.1
psutil==5.9.8
pydantic==2.10.6
pydantic_core==2.27.2
sniffio==1.3.1
starlette==0.45.3
tabulate==0.9.0
typing-inspect==0.9.0
typing_extensions==4.12.2
uvicorn==0.34.0
yarl==1.18.3
bdfparser==2.2.0
charset-normalizer==3.4.1
libsixel-python==0.5.0
numpy==2.2.2
pillow==11.1.0
pygame==2.6.1
requests==2.32.3
RGBMatrixEmulator==0.12.1
tornado==6.4.2
urllib3==2.3.0
pytz==2025.1
python-dotenv==1.0.1